Fix zsh prompt in bare repositories
The PROMPT uses git status to check how many uncommitted changes there
are in the project, however this command fails in bare repositories. So
we now wrap those checks in a ! core.bare to make sure that these checks
are not performed if you happen to be inside a bare repo at the time.