Also support shift + lowering the volume
While this is not strictly required before, this is useful as it allows
us to scroll up and down with the shift key pressed. As this has been
explictly added now, we've also added the cap to the non-shift scroll
down behaviour. So scrolling the volume in either direction will now
auto bring it within the cap.