Amend the logic for displaying fastfetch
Previously, we would only show fastfetch if the window was detected as
being right up against the edge of the screen (less than 100 pixels from
the edge).

This logic is now improved so the window does not need to be right up at
the edge, it merely needs to be the top left-most window.

The new implementation also no longer expects the window to be on the
active workplace, meaning the correct logic can be used even if a window
is spawned in the background on another workplace.